# src/document/swagger-final.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sekolah Kopi Raisa",
    "version": "1.0.0",
    "description": "API untuk aplikasi Sekolah Kopi Raisa - E-commerce dan Management System"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    },
    {
      "url": "https://be-web-sekolah-kopi-raisa.vercel.app",
      "description": "Vercel Production Server"
    },
    {
      "url": "https://sekolah-kopi-raisa.up.railway.app",
      "description": "Railway Production Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token"
      }
    }
  },
  "paths": {
...
    "/api/v1/product": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "Products retrieved successfully"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create new product",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "integer"
                  },
                  "productFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "partner_id": {
                    "type": "integer"
                  },
                  "weight": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully"
          }
        }
      }
    },
    "/api/v1/product/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product retrieved successfully"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update product",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "stock": {
                    "type": "integer"
                  },
                  "productFile": {
                    "type": "string",
                    "format": "binary"
                  },
                  "partner_id": {
                    "type": "number"
                  },
                  "weight": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Product updated successfully"
          }
        }
      }
    },
    "/api/v1/product/{idProduct}": {
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete product",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "idProduct",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deleted successfully"
          }
        }
      }
    },
...
  }
}